==> 1_Spec_Preprocessing.ipynb <==
# 2022 빅콘테스트_spec 데이터 전처리
## 우 걱정마세요

## 1. 라이브러리 및 패키지 불러오기

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

user_spec = pd.read_csv('/Users/sungahwang/Desktop/BC_data/user_spec.csv')
user_spec.head()

loan_result = pd.read_csv('/Users/sungahwang/Desktop/BC_data/loan_result.csv')
loan_result.head()

## 2. Loan_result에 없는 id 제거

# user_spec 데이터의 현재 결측값 확인
user_spec.isnull().sum()

respec = user_spec
drop_application_id = list(set(respec['application_id'])-set(loan_result['application_id']))

respec.drop(respec[respec['application_id'].isin(drop_application_id)].index,inplace=True)

# 결측값이 잘 채워졌는지 확인
respec.isnull().sum()

## 3. yearly_income 컬럼

# yaerly_income 컬럼의 1개의 결측값은 무직이라는 판단하에 0으로 결측값을 채워줌.
respec.loc[respec['yearly_income'].isnull(), 'yearly_income'] = 0

## 4. purpose 컬럼

respec['purpose'].unique()

# 영어와 한글을 통일
respec['purpose'] = respec['purpose'].replace("ETC", '기타').replace("SWITCHLOAN", '대환대출').replace("LIVING", '생활비').replace("BUSINESS", "사업자금").replace("BUYHOUSE", '주택구입').replace("HOUSEDEPOSIT", '전월세보증금').replace("INVEST", '투자').replace("BUYCAR", '자동차구입')


respec.groupby(['purpose']).mean()

## 5. birth_year 컬럼
* birth_year을 나이 파생 변수를 만들어 대체
* 나이를 10단위로 바꿔주었고, 70이상의 값은 70으로 처리
* yearly_income을 기준으로 null값을 채움

# 성별이 다른 신청서에는 있지만 결측치가 있는 유저들 처리
gender_null = set(respec[(respec['user_id'].isin(set(respec[respec['gender'].isnull()]['user_id'])))&(respec['gender'].notnull())]['user_id'])

for i in gender_null:
    respec.loc[(respec['user_id']==i)&(respec['birth_year'].isnull()),'birth_year'] = respec[respec['user_id'] == i]['birth_year'].dropna().unique().item()
    respec.loc[(respec['user_id']==i)&(respec['gender'].isnull()),'gender'] = respec[respec['user_id'] == i]['gender'].dropna().unique().item()

# 나이 파생변수 추가
respec['age'] = 2023 - respec['birth_year']

# 나이를 10단위로 바꿔주고, 70이상의 값은 70으로 처리
respec.loc[respec['age'] < 20, 'reage'] = 10
respec.loc[(respec['age'] >= 20) & (respec['age'] <30), 'reage'] = 20
respec.loc[(respec['age'] >= 30) & (respec['age'] <40), 'reage'] = 30
respec.loc[(respec['age'] >= 40) & (respec['age'] <50), 'reage'] = 40
respec.loc[(respec['age'] >= 50) & (respec['age'] <60), 'reage'] = 50
respec.loc[(respec['age'] >= 60) & (respec['age'] <70), 'reage'] = 60
respec.loc[respec['age'] >= 70, 'reage'] = 70

# 나이에 따른 평균 값들 확인
respec_group = respec.groupby(['reage']).mean()
respec_group = respec_group.reset_index()

respec_group

#나이에 따라 소득이 다르다는 판단하에 나이의 결측값을 소득이 속하는 평균으로 채워줌.
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] < respec_group[respec_group['reage'] == 20]['yearly_income'].item()), 'reage'] = 10
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] >= respec_group[respec_group['reage'] == 20]['yearly_income'].item()) & (respec['yearly_income'] < respec_group[respec_group['reage'] == 70]['yearly_income'].item()), 'reage'] = 20
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] >= respec_group[respec_group['reage'] == 70]['yearly_income'].item()) & (respec['yearly_income'] < respec_group[respec_group['reage'] == 30]['yearly_income'].item()), 'reage'] = 70
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] >= respec_group[respec_group['reage'] == 30]['yearly_income'].item()) & (respec['yearly_income'] < respec_group[respec_group['reage'] == 60]['yearly_income'].item()), 'reage'] = 30
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] >= respec_group[respec_group['reage'] == 60]['yearly_income'].item()) & (respec['yearly_income'] < respec_group[respec_group['reage'] == 40]['yearly_income'].item()	), 'reage'] = 60
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] >= respec_group[respec_group['reage'] == 40]['yearly_income'].item()) & (respec['yearly_income'] < respec_group[respec_group['reage'] == 50]['yearly_income'].item()), 'reage'] = 40
respec.loc[(respec['reage'].isnull()) & (respec['yearly_income'] >= respec_group[respec_group['reage'] == 50]['yearly_income'].item()) , 'reage'] = 50

respec = respec.drop(['birth_year', 'age'], axis = 1)

## 6. gender 컬럼
- income_type별 gender의 비율을 구하여 비율에 맞춰 결측값을 채워줌

respec['income_type'].unique()

respec['income_type'] = respec['income_type'].replace("PRIVATEBUSINESS", 1).replace("FREELANCER", 2).replace("EARNEDINCOME", 3).replace("OTHERINCOME", 4).replace("EARNEDINCOME2", 5).replace("PRACTITIONER", 6)

sample1 = respec[(respec['gender'].notnull()) & (respec['income_type'] == 1)]['gender'].value_counts()

per_0 = sample1[0] / (sample1[1] + sample1[0])
per_0 = (per_0*100).round(0)

# income_type == 1
rnd_num = np.random.uniform(0, 1, 1)

if rnd_num > per_0:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 1), 'gender'] = 1
else:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 1), 'gender'] = 0

sample1 = respec[(respec['gender'].notnull()) & (respec['income_type'] == 2)]['gender'].value_counts()

per_0 = sample1[0] / (sample1[1] + sample1[0])
per_0 = (per_0*100).round(0)

# income_type == 2
rnd_num = np.random.uniform(0, 1, 1)

if rnd_num > per_0:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 2), 'gender'] = 1
else:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 2), 'gender'] = 0

sample1 = respec[(respec['gender'].notnull()) & (respec['income_type'] == 3)]['gender'].value_counts()

per_0 = sample1[0] / (sample1[1] + sample1[0])
per_0 = (per_0*100).round(0)

# income_type == 3
rnd_num = np.random.uniform(0, 1, 1)

if rnd_num > per_0:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 3), 'gender'] = 1
else:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 3), 'gender'] = 0

sample1 = respec[(respec['gender'].notnull()) & (respec['income_type'] == 4)]['gender'].value_counts()

per_0 = sample1[0] / (sample1[1] + sample1[0])
per_0 = (per_0*100).round(0)

# income_type == 4
rnd_num = np.random.uniform(0, 1, 1)

if rnd_num > per_0:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 4), 'gender'] = 1
else:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 4), 'gender'] = 0

sample1 = respec[(respec['gender'].notnull()) & (respec['income_type'] == 5)]['gender'].value_counts()

per_0 = sample1[0] / (sample1[1] + sample1[0])
per_0 = (per_0*100).round(0)

# income_type == 5
rnd_num = np.random.uniform(0, 1, 1)

if rnd_num > per_0:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 5), 'gender'] = 1
else:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 5), 'gender'] = 0

sample1 = respec[(respec['gender'].notnull()) & (respec['income_type'] == 6)]['gender'].value_counts()

per_0 = sample1[0] / (sample1[1] + sample1[0])
per_0 = (per_0*100).round(0)

# income_type == 6
rnd_num = np.random.uniform(0, 1, 1)

if rnd_num > per_0:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 6), 'gender'] = 1
else:
  respec.loc[(respec['gender'].isnull()) & (respec['income_type'] == 6), 'gender'] = 0

## 7. company_enter_month 컬럼
- 년도를 쓰는 기준이 다른것을 통일해줌(2022년을 2022,22라고 표기되어 있음)


# 표기 방법 통일
respec.loc[(respec['company_enter_month'] < 10000000),'company_enter_month'] = respec['company_enter_month'] / 100

respec.loc[(respec['company_enter_month'] >= 10000000),'company_enter_month'] = respec['company_enter_month'] / 10000

respec['company_enter_month'] = respec['company_enter_month'].round()

sample = respec[respec['company_enter_month']<0]

respec.groupby(['reage'])['company_enter_month'].mean()

# 입사년도의 결측값을 나이를 기준으로 채워줌.
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] <= 10),'company_enter_month'] = 2020
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] > 10) & (respec['reage'] <= 20),'company_enter_month'] = 2021
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] > 20) & (respec['reage'] <= 30),'company_enter_month'] = 2020
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] > 30) & (respec['reage'] <= 40),'company_enter_month'] = 2017
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] > 40) & (respec['reage'] <= 50),'company_enter_month'] = 2016
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] > 50) & (respec['reage'] <= 60),'company_enter_month'] = 2017
respec.loc[(respec['company_enter_month'].isnull()) & (respec['reage'] > 60),'company_enter_month'] = 2018

# 입사년으로 커리어 파생 변수 생성
respec['career'] = 2022 - respec['company_enter_month'] + 1

respec = respec.drop(['company_enter_month'], axis = 1)

## 8. existing_loan_cnt

# 데이터 셋 내에 기대출수가 0인 사람이 없으므로, NULL 값인 user의 기대출수는 0일 것이라 판단
respec.loc[respec['existing_loan_cnt'].isnull(), 'existing_loan_cnt'] = 0

## 9. existing_loan_amt

# 기 대출수가 0이면 기대출 금액도 0(오류 처리)
respec.loc[(respec['existing_loan_amt'].isnull()) & (respec['existing_loan_cnt']==0), 'existing_loan_amt'] = 0

print(respec['existing_loan_amt'].describe())

# 1사분면 ~ 3사분면까지의 값의 평규능로 기대출금액의 결측값을 채워줌
respec.loc[(respec['existing_loan_amt'].isnull()), 'existing_loan_amt'] = (7.700000e+07 + 1.000000e+07) /2

## 10. personal_rehabilitation_yn

pry_400 =respec[respec['credit_score']<=400]
pry_400.head()

respec['personal_rehabilitation_yn'] = pry_400['personal_rehabilitation_yn'].fillna(1)

respec['personal_rehabilitation_yn'].isnull().sum()

df=respec[respec['personal_rehabilitation_yn'].isnull()]
df_sample = df.sample(n=569011, random_state=100)
df_sample

respec['personal_rehabilitation_yn'] = df_sample['personal_rehabilitation_yn'].fillna(0)

respec['personal_rehabilitation_yn'].isnull().sum()

respec['personal_rehabilitation_yn'] = respec['personal_rehabilitation_yn'].fillna(1)

respec['personal_rehabilitation_yn'].isnull().sum()

## 11. personal_rehabilitation_complete_yn

pr_0 = respec['personal_rehabilitation_yn'] == 0
prc_0 = respec['personal_rehabilitation_complete_yn'] == 0

subset = respec[pr_0 & prc_0]
subset

subset['personal_rehabilitation_complete_yn'] =subset['personal_rehabilitation_complete_yn'].replace(0, 2)
subset

pr_1 = respec['personal_rehabilitation_yn'] != 0
prc_1 = respec['personal_rehabilitation_complete_yn'] != 0

subset1 = respec[pr_1 | prc_1]

prcy_600 =respec[respec['credit_score']>=600]
prcy_600.head()

respec['personal_rehabilitation_complete_yn'] = prcy_600['personal_rehabilitation_complete_yn'].fillna(1)

respec['personal_rehabilitation_complete_yn'] = respec['personal_rehabilitation_complete_yn'].fillna(0)

respec['income_type'] = respec['income_type'].replace(1, '개인사업자').replace(2, '프리랜서').replace(3, '직장가입자(4대보험O)').replace(4, '기타소득').replace(5, '직장가입자(4대보험X)').replace(6, '전문직')

respec.head()

## 12. 모델링

from sklearn.preprocessing import LabelEncoder # 라벨 인코더
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.metrics import roc_auc_score, roc_curve, precision_recall_curve, precision_score, recall_score, auc
from sklearn.metrics import confusion_matrix, plot_confusion_matrix
from sklearn.metrics import classification_report, accuracy_score, f1_score
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.preprocessing import StandardScaler

respec1 = respec.drop(['application_id', 'user_id', 'insert_time'], axis = 1)

respec_train = respec1[respec1['credit_score'].notnull()]
respec_test = respec1[respec1['credit_score'].isnull()]

# 인코딩
respec_train = pd.get_dummies(respec_train)
respec_test= pd.get_dummies(respec_test)

train_columns = respec_train.columns
test_columns = respec_test.columns

respec_train = respec_train.values.astype(float)
respec_test = respec_test.values.astype(float)

respec_train = pd.DataFrame(respec_train, columns = train_columns)
respec_test = pd.DataFrame(respec_test, columns = test_columns)

data = respec_train.drop('credit_score',axis=1)
target = respec_train['credit_score']

standardScaler = StandardScaler()
print(standardScaler.fit(data))
data_scaled = standardScaler.transform(data)

from sklearn.model_selection import train_test_split
X_train, X_val, y_train, y_val = train_test_split(data_scaled, target, test_size=0.3, random_state=0)

### 테스트

test_data = respec_test.drop('credit_score',axis=1)
target = respec_test['credit_score']

standardScaler = StandardScaler()
print(standardScaler.fit(test_data))
data_scaled = standardScaler.transform(test_data)

### XGB

import xgboost

model = xgboost.XGBRegressor(n_estimators=1000, learning_rate=0.2, max_depth=14, random_state = 2022, colsample_bytree = 0.75, min_child_weight = 0.5)
model.fit(X_train, y_train)

y_pred = model.predict(X_val)

mse = mean_squared_error(y_val, y_pred)
rmse = np.sqrt(mse)
print('MSE : {0:.3f}, RMSE : {1:.3F}'.format(mse , rmse))
print('Variance score : {0:.3f}'.format(r2_score(y_val,y_pred)))

xbg_preds_test = model.predict(test_data.to_numpy())

respec.loc[(respec['credit_score'].isnull()), 'credit_score'] = xbg_preds_test.round()

### 전처리 및 EDA


# KCB 신용등급 점수표 기준
respec.loc[(respec['credit_score'] >= 942) & (respec['credit_score'] <= 1000),'credit_grade'] = 1
respec.loc[(respec['credit_score'] >= 891) & (respec['credit_score'] <= 941),'credit_grade'] = 2
respec.loc[(respec['credit_score'] >= 832) & (respec['credit_score'] <= 890),'credit_grade'] = 3
respec.loc[(respec['credit_score'] >= 768) & (respec['credit_score'] <= 831),'credit_grade'] = 4
respec.loc[(respec['credit_score'] >= 698) & (respec['credit_score'] <= 767),'credit_grade'] = 5
respec.loc[(respec['credit_score'] >= 630) & (respec['credit_score'] <= 697),'credit_grade'] = 6
respec.loc[(respec['credit_score'] >= 530) & (respec['credit_score'] <= 629),'credit_grade'] = 7
respec.loc[(respec['credit_score'] >= 454) & (respec['credit_score'] <= 529),'credit_grade'] = 8
respec.loc[(respec['credit_score'] >= 335) & (respec['credit_score'] <= 453),'credit_grade'] = 9
respec.loc[(respec['credit_score'] >= 0) & (respec['credit_score'] <= 334),'credit_grade'] = 10

respec = respec.drop(['credit_score'], axis = 1)

respec.isnull().sum()

## 13. 전처리한 spec데이터 군집화(Cluster:5)

from sklearn.preprocessing import PowerTransformer
from kmodes.kprototypes import KPrototypes

data = respec.drop(columns=['application_id','user_id','insert_time'])

kproto = KPrototypes(n_clusters = 5, init='Huang',n_jobs=-1,n_init=1,verbose=True)
kproto.fit(data, categorical=[0,2,3,4,6,7,8])
centroids = kproto.cluster_centroids_
labels  = kproto.labels_

respec['spec_clust'] = labels

respec.to_csv("/Users/sungahwang/Desktop/BC_data/respec_final.csv", index = False)
