==> 3_Preprocessing_Log+Spec+Loan.ipynb <==
# 1. 라이브러리 및 패키지 불러오기

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

# 2. 데이터 불러오기

log_data = pd.read_csv('/Users/sungahwang/Desktop/BC_data/log_data.csv')

relog = pd.read_csv('/Users/sungahwang/Desktop/BC_data/relog.csv')

relog = relog.drop(['Unnamed: 0'], axis = 1)

# 3. 전처리

##1) mp_os
- android os는 getcreditinfo행동밖에 없고, Android os는 getcreditinfo만 없으므로 합쳐주는게 맞다고 판단 android -> Android로 변환

- 애플 제품에는 login이 없음 / 아이패드에는 여윳돈 계산 기능이 없는것인지 하지 않은 것인지 암튼 없음

log_data.loc[list(log_data[log_data['mp_os']=='android'].index),'mp_os'] = 'Android'

log_data.loc[list(log_data[log_data['mp_os']=='iPadOS'].index),'mp_os'] = 'iOS'

# mp_os가 결측값이면 mp_app_version도 결측값임을 확인

# 4. 파생변수 추가

## 1) 각 유저가 하루에 행동을 얼마나 했는가?

log = log_data

grouped = log.groupby(['user_id', 'date_cd'])['user_id'].count()
df_grouped = pd.DataFrame(grouped)

df_grouped.rename(columns = {'user_id' : 'user_id_cnt'}, inplace = True)

df_grouped_new = df_grouped.reset_index('date_cd')
df_grouped_new = df_grouped_new.reset_index('user_id')

df_grouped_new.head()

## 2) 유저가 했던 행동을 총합해서 하나의 컬럼으로 추출

sample = pd.DataFrame(df_grouped_new['user_id_cnt'].groupby(df_grouped_new['user_id']).sum())

log = pd.merge(relog , sample,
                 how = 'left',
                 on = 'user_id')
log

## 3) 유저가 기간동안 총 몇일 접속했는지

grouped = log_data.groupby(['user_id', 'date_cd'])['user_id'].count()
df_grouped = pd.DataFrame(grouped)

df_grouped.rename(columns = {'user_id' : 'user_id_day_cnt'}, inplace = True)
df_grouped_new = df_grouped.reset_index('date_cd')
df_grouped_new = df_grouped_new.reset_index('user_id')
df_grouped_new.head()

df_grouped_new['user_id_day_cnt'].groupby(df_grouped_new['user_id']).count()

sample1=pd.DataFrame(df_grouped_new['user_id_day_cnt'].groupby(df_grouped_new['user_id']).count())
sample1.head()

log = pd.merge(log , sample1,
                 how = 'left',
                 on = 'user_id')
log.head()

## 4) 한도 조회, 신용 정보 조회, 대출 관리 서비스 이용, 여윳돈 계산기 서비스 이용, DSR 계산기 서비스 이용 중 어떤 것을 먼저 이용한 사람인가에 따른 차이점

log1 = log_data

group = log1.groupby(['user_id', 'event'])['timestamp'].min()
df_group = pd.DataFrame(group)

df_group_new = df_group.reset_index()
df_final = df_group_new.drop_duplicates(['user_id'], keep = 'first')

df_final['event'].unique()

df_final.loc[(df_final['event'] != 'GetCreditInfo') & (df_final['event'] != 'EndLoanApply')& 
             (df_final['event'] != 'UseLoanManage') & (df_final['event'] != 'ViewLoanApplyIntro')&
             (df_final['event'] != 'StartLoanApply') & (df_final['event'] != 'CompleteIDCertification'), 'event'] = 'others'

df_final = df_final.rename({'event':'first_event'}, axis = 1)
df_final

df_final = df_final.drop(['timestamp'], axis = 1)

log = pd.merge(log , df_final,
                 how = 'left',
                 on = 'user_id')
log.head()

# 5. Timeout 10분을 기준으로 세션 나누기

log2 = log_data
log2 = log2.sort_values(["user_id", "timestamp"], ascending= True)

log2['timestamp'] = pd.to_datetime(log2['timestamp'])
log2.head()

log2 = log2.drop(['mp_os', 'mp_app_version', 'date_cd'], axis = 1)

log2['timeout'] = log2.timestamp - log2.timestamp.shift()
log2.head()

pastid = 0
userid = 0

# timeout의 기준을 10으로 함.
for i in range(len(log2)):
  userid = log2['user_id'].values[i]
  if userid != pastid:
    log2['timeout'].values[i] = 0
  if  log2['timeout'].values[i] > 10000000000:
    log2['timeout'].values[i] = 0
  pastid = userid

log2.head(8)

grouped = log2.groupby(['user_id'])['timeout'].sum()
df_grouped = pd.DataFrame(grouped)

df_grouped

df_grouped.rename(columns = {'timeout' : 'timeout_sum'}, inplace = True)

log = pd.merge(log , df_grouped,
                 how = 'left',
                 on = 'user_id')
log.head()

## timeout 초 단위

log['timeout_sum(s)'] = pd.to_timedelta(log['timeout_sum']).view(np.int64) / 1e9
log = log.drop(['timeout_sum'], axis = 1)

log

## 7. Log+Spec+Loan

join = pd.read_csv('/Users/sungahwang/Desktop/BC_data/Preprocessing+loan.csv')

join = join.rename({'Labels':'spec_clust'}, axis = 1)

log.rename(columns={'user_id_cnt': 'action_cnt'}, inplace=True)

log.rename(columns={'user_id_day_cnt': 'use_day_cnt'}, inplace=True)

# first_event 결측값 -> Others
log['first_event'] = log['first_event'].fillna('Others')

join_final = pd.merge(join,log ,on="user_id",how="left")

## 전처리

# id 당 총 대출 수
id_applied = join_final.groupby('user_id').sum()['is_applied'].sort_values(ascending=False)
id_applied = pd.DataFrame(id_applied)
id_applied = id_applied.reset_index()
id_applied = id_applied.rename({'is_applied':'sum_applied'}, axis = 1)

join = pd.merge(id_applied,join_final ,on="user_id",how="right")

sum_applied_list = id_applied['sum_applied'].unique()

## 8. 결측값

### Join으로 인하여 생기는 결측값 중 is_applied가 0인 데이터는 제거

join.isnull().sum()

join_0 = join[join['is_applied']==0]
join_not0 = join[join['is_applied']!=0]

join_0 = join_0.dropna(subset = ['timestamp'], axis=0) 

join = pd.concat([join_0,join_not0])

join = join.drop(['timestamp'], axis = 1)

join = join.drop(['event'], axis = 1)

from pandas import Series, DataFrame

for i in range(len(sum_applied_list)):
  join.loc[join['action_cnt'].isnull() & join['sum_applied']== sum_applied_list[i], 'action_cnt'] = join[(join['sum_applied']==sum_applied_list[i])]['action_cnt'].mean()
  join.loc[join['use_day_cnt'].isnull() & join['sum_applied']== sum_applied_list[i], 'use_day_cnt'] = join[(join['sum_applied']==sum_applied_list[i])]['use_day_cnt'].mean()
  join.loc[join['timeout_sum(s)'].isnull() & join['sum_applied']== sum_applied_list[i], 'timeout_sum(s)'] = join[(join['sum_applied']==sum_applied_list[i])]['timeout_sum(s)'].mean()

df_max = join.groupby(['cluster2', 'first_event'])['first_event'].value_counts()
df_max

join.loc[join['first_event'].isnull(), 'first_event'] = 'CompleteIDCertification'

join.isnull().sum()

join.to_csv('/Users/sungahwang/Desktop/BC_data/final.csv', index = False)
