==> 4_Finalmodel.ipynb <==
#pip install bayesian-optimization

#pip install catboost

#pip install imblearn

# 1. 라이브러리 및 전처리

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

# 패키지 설치
import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns

from sklearn.model_selection import cross_val_score
from sklearn.metrics import roc_auc_score, roc_curve, precision_recall_curve, precision_score, recall_score, auc
from sklearn.metrics import confusion_matrix, plot_confusion_matrix
from sklearn.metrics import classification_report, accuracy_score, f1_score

from bayes_opt import BayesianOptimization

from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
import warnings
warnings.filterwarnings('ignore')

from imblearn.under_sampling import RandomUnderSampler
from collections import Counter
from imblearn.over_sampling import RandomOverSampler

final = pd.read_csv("C:\\Users\\user\\Desktop\\빅콘테스트\\final.csv")

final_o = pd.read_csv("C:\\Users\\user\\Desktop\\빅콘테스트\\final.csv")

result = pd.read_csv("C:\\Users\\user\\Desktop\\빅콘테스트\\데이터분석분야_퓨처스부문_평가데이터.csv")

final = final.drop(['user_id', 'application_id', 'insert_time', 'loanapply_insert_time'], axis = 1)

final['income_type'] = final['income_type'].astype('category')
final['employment_type'] = final['employment_type'].astype('category')
final['houseown_type'] = final['houseown_type'].astype('category')
final['purpose'] = final['purpose'].astype('category')
final['first_event'] = final['first_event'].astype('category')
final['weekday'] = final['weekday'].astype('category')
final['personal_rehabilitation_yn'] = final['personal_rehabilitation_yn'].astype('category')
final['personal_rehabilitation_complete_yn'] = final['personal_rehabilitation_complete_yn'].astype('category')
final['gender'] = final['gender'].astype('category')
final['spec_clust'] = final['spec_clust'].astype('category')
final['cluster2'] = final['cluster2'].astype('category')

final['yearly_income']=np.log(final['yearly_income'])
final["desired_amount"]=np.log(final["desired_amount"])
final["existing_loan_amt"]=np.log(final["existing_loan_amt"])

final = pd.get_dummies(final)

final_train = final[final['is_applied'].notnull()]
final_test = final[final['is_applied'].isnull()]

data = final_train.drop('is_applied',axis=1)
target = final_train['is_applied']

from sklearn.model_selection import train_test_split
X_train, X_val, y_train, y_val = train_test_split(data, target, test_size=0.3, random_state=0)

test_data = final_test.drop('is_applied',axis=1)
target = final_test['is_applied']

rus = RandomUnderSampler(random_state= 42, sampling_strategy= {0:600000, 1:392199 })

X_rus, y_rus = rus.fit_resample(X_train, y_train)
print(Counter(y_rus))

### catboost / sclae_pos_weight

cb_rus = CatBoostClassifier(n_estimators=1000, 
                            learning_rate=0.3, 
                            max_depth=10, 
                            scale_pos_weight = 5,
                            random_state = 42)
cb_rus.fit(X_train, y_train)
cb_rus_preds= cb_rus.predict(X_val)

#f1 score
print('f1 score : {:0.4f}'.format(f1_score(cb_rus_preds, y_val)))
#정확도
print('accuracy: {:0.4f}'.format(accuracy_score(cb_rus_preds, y_val)))
# 정밀도
print('precision: {:0.4f}'.format(precision_score(cb_rus_preds, y_val)))
#재현율
print('recall: {:0.4f}'.format(recall_score(cb_rus_preds, y_val)))
# ROC_AUC
print('roc_auc: {:0.4f}'.format(roc_auc_score(cb_rus_preds, y_val)))

cat_preds_test = cb_rus.predict(test_data.to_numpy())

final_o

result_final = final_o[final_o['is_applied'].isnull()]

result_final['is_applied'] = cat_preds_test

result_final = result_final.drop_duplicates(['application_id', 'product_id'])

drop_application_id = list(set(result_final['application_id'])-set(result['application_id']))

result_final.drop(result_final[result_final['application_id'].isin(drop_application_id)].index,inplace=True)

result_final = result_final[['application_id','product_id','is_applied']]

result_final = result_final.sort_values(by=['application_id'],ascending = True)

sample = pd.merge(result_final,result,on=["application_id",'product_id'],how="right")

sample.to_csv('C:\\Users\\user\\Desktop\\빅콘테스트\\result_final.csv', index = False)
